Updated formatting methods for Formatter.java to assist TextUI.java and practice SRP
