+++ src/seedu/addressbook/ui/Formatter.java
@@ -1,5 +1,7 @@
 package seedu.addressbook.ui;
 
+import java.util.List;
+
 public class Formatter {
     /** A decorative prefix added to the beginning of lines printed by AddressBook */
     public static final String LINE_PREFIX = "|| ";
@@ -18,5 +20,24 @@ public class Formatter {
     /** Format of a comment input line. Comment lines are silently consumed when reading user input. */
     public static final String COMMENT_LINE_FORMAT_REGEX = "#.*";
 
+    /** Formats a list of strings as a viewable indexed list. */
+    public static String getIndexedListForViewing(List<String> listItems) {
+        final StringBuilder formatted = new StringBuilder();
+        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
+        for (String listItem : listItems) {
+            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
+            displayIndex++;
+        }
+        return formatted.toString();
+    }
+
+    /**
+     * Formats a string as a viewable indexed list item.
+     *
+     * @param visibleIndex visible index for this listing
+     */
+    private static String getIndexedListItem(int visibleIndex, String listItem) {
+        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
+    }
 
 }
